deploy/requirements.txt
jinja2
docopt
termcolor
pyyaml
